backend-jo/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic[email]
python-jose[cryptography]
